requirements.txt
PyJWT>=2.8.0
click
python-dotenv
pyyaml

